requirements.txt
dj-database-url==0.4.1
Django==1.11
gunicorn==19.6.0
psycopg2==2.6.2
whitenoise==2.0.6
python-twitter==3.2.1
